Refactor and add Regions
