+++ src/Classes/UpdateMessage.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Certify.Classes
+{
+    public class UpdateMessage
+    {
+        public string Body { get; set; }
+        public string DownloadPageURL { get; set; }
+        public string ReleaseNotesURL { get; set; }
+    }
+}
