+++ src/Forms/Controls/CertRequestSettingsIIS.cs
@@ -11,6 +11,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using Certify.Classes;
 
 namespace Certify.Forms.Controls
 {
