+++ src/Management/Util.cs
@@ -4,82 +4,10 @@ using System.Linq;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
+using Certify.Classes;
 
 namespace Certify.Management
 {
-    public class AppVersion
-    {
-        public int Major { get; set; }
-        public int Minor { get; set; }
-        public int Patch { get; set; }
-
-        public static AppVersion FromString(string version)
-        {
-            string[] versionComponents = version.Split('.');
-
-            AppVersion current = new AppVersion
-            {
-                Major = int.Parse(versionComponents[0]),
-                Minor = int.Parse(versionComponents[1]),
-                Patch = int.Parse(versionComponents[2])
-            };
-            return current;
-        }
-
-        public static bool IsOtherVersionNewer(AppVersion currentVersion, AppVersion otherVersion)
-        {
-            if (currentVersion.Major >= otherVersion.Major)
-            {
-                if (currentVersion.Major > otherVersion.Major)
-                {
-                    return false;
-                }
-
-                //current major version is same, check minor
-                if (currentVersion.Minor >= otherVersion.Minor)
-                {
-                    if (currentVersion.Patch < otherVersion.Patch)
-                    {
-                        return true;
-                    }
-                    else
-                    {
-                        return false;
-                    }
-                }
-
-                //current minor version is less
-                if (currentVersion.Minor < otherVersion.Minor)
-                {
-                    return true;
-                }
-            }
-            else
-            {
-                //other Major version is newer
-                return true;
-            }
-
-            return false; ;
-        }
-    }
-
-    public class UpdateMessage
-    {
-        public string Body { get; set; }
-        public string DownloadPageURL { get; set; }
-        public string ReleaseNotesURL { get; set; }
-    }
-
-    public class UpdateCheck
-    {
-        public AppVersion Version { get; set; }
-
-        public UpdateMessage Message { get; set; }
-
-        public bool IsNewerVersion { get; set; }
-    }
-
     public class Util
     {
         public async Task<UpdateCheck> CheckForUpdates(string appVersion)
@@ -116,5 +44,6 @@ namespace Certify.Management
                 return null;
             }
         }
+
     }
 }
